sysctl takes a 'size_t' variable for the size parameter
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-5-patches@19829 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/host/data_access/swrun_kinfo.c
@@ -110,7 +110,7 @@ netsnmp_arch_swrun_init(void)
 {
 #if NETSNMP_CAN_USE_SYSCTL && defined(CTL_KERN) && defined(KERN_MAXPROC)
     extern int _swrun_max;
-    int max_size = sizeof(_swrun_max);
+    size_t max_size = sizeof(_swrun_max);
     int maxproc_mib[] = { CTL_KERN, KERN_MAXPROC };
     sysctl(maxproc_mib, 2, &_swrun_max, &max_size, NULL, 0);
 #endif
